Cleanup, style fixes
