+++ Source/Core/Subtitles/Subtitles.h
@@ -1,6 +1,12 @@
+// Copyright 2023 Dolphin Emulator Project
+// SPDX-License-Identifier: GPL-2.0-or-later
+
 #pragma once
+
 #include <string>
-#include <DiscIO\Filesystem.h>
+
+#include "DiscIO/Filesystem.h"
+
 namespace Subtitles
 {
 const std::string BottomOSDStackName = "subtitles-bottom";
