+++ Source/Core/Subtitles/WebColors.h
@@ -1,7 +1,12 @@
+// Copyright 2023 Dolphin Emulator Project
+// SPDX-License-Identifier: GPL-2.0-or-later
+
 #pragma once
-#include <Common/CommonTypes.h>
+
 #include <map>
 #include <string>
+
+#include "Common/CommonTypes.h"
 
 namespace Subtitles
 {
